[refactor] Move LLVM CustomType-related functions to codegen_llvm_quant.cpp (#2173)
